update alias on its own component is completed
